backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
google-genai>=0.2.0
python-dotenv>=1.0.0